user authentication

+++ app/routers/post.py
@@ -1,8 +1,8 @@
 from fastapi import FastAPI, Depends, HTTPException, status,APIRouter,Response
 from sqlalchemy.orm import Session
-from .. import models,schemas
+from .. import models,schemas,oauth2
 from ..database import get_db
-from typing import List
+from typing import List,Optional
 
 
 
@@ -14,49 +14,53 @@ router = APIRouter(
 
 
 @router.get("/",response_model=List[schemas.PostResponse])
-def get_posts(db:Session = Depends(get_db)):
+def get_posts(db:Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user),limit:int=10,skip:int=0,search:Optional[str]=""):
     # cursor.execute("""SELECT * FROM posts""")
     # posts = cursor.fetchall()
-    posts = db.query(models.Post).all()
+    posts = db.query(models.Post).filter(models.Post.owner_id == current_user.id).filter(models.Post.title.contains(search)).limit(limit).offset(skip).all()
     return posts
 
 @router.post("/",response_model=schemas.PostResponse,status_code=status.HTTP_201_CREATED)
-async def create_posts(post:schemas.PostCreate,db:Session = Depends(get_db)):
+async def create_posts(post:schemas.PostCreate,db:Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
     # cursor.execute(""" INSERT INTO posts (title,content,published) VALUES (%s,%s,%s) RETURNING *""",
     # (post.title,post.content,post.published))
     # new_post = cursor.fetchone()
     # conn.commit()
-    new_post = models.Post(**post.dict())
+    new_post = models.Post(**post.dict(),owner_id=current_user.id)
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
     return new_post
 
 @router.get("/{id}",response_model=schemas.PostResponse)
-def get_post(id:int,db:Session = Depends(get_db)):
+def get_post(id:int,db:Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
     # cursor.execute(""" SELECT * FROM posts WHERE id = %s """,(id,))
     # post = cursor.fetchone()
     post = db.query(models.Post).filter(models.Post.id == id).first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail=f"post with id {id} not found")
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,detail="You are not authorized to perform requested action")
     return post
 
 
 @router.delete("/{id}",status_code=status.HTTP_204_NO_CONTENT)
-def delete_post(id:int,db:Session = Depends(get_db)):
+def delete_post(id:int,db:Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
     # cursor.execute("""DELETE FROM posts WHERE id = %s RETURNING *""",(id,))
     # deleted_post = cursor.fetchone()
     # conn.commit()
     post = db.query(models.Post).filter(models.Post.id == id).first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail=f"post with id {id} not found")
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,detail="You are not authorized to perform requested action")
     db.delete(post)
     db.commit()
     return Response( status_code=status.HTTP_204_NO_CONTENT)
 
 
 @router.put("/{id}",response_model=schemas.PostResponse)
-def update_post(id:int,updated_post:schemas.PostCreate,db:Session = Depends(get_db)):
+def update_post(id:int,updated_post:schemas.PostCreate,db:Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
     # cursor.execute("""UPDATE posts SET title = %s, content = %s, published = %s WHERE id = %s RETURNING *""",
     # (post.title,post.content,post.published,id))
     # updated_post = cursor.fetchone()
@@ -65,7 +69,8 @@ def update_post(id:int,updated_post:schemas.PostCreate,db:Session = Depends(get_
     post = post_query.first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail=f"post with id {id} not found")
-        
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,detail="You are not authorized to perform requested action")
     post_query.update(updated_post.dict(),synchronize_session=False)
     db.commit()
     return post_query.first()
